requirements.txt
streamlit==1.25.0
SQLAlchemy==2.0.22
reportlab==3.6.12
python-dotenv==1.0.0
bcrypt==4.0.1
pillow==9.5.0
pandas==2.1.1
